requirements.txt
numpy
Pillow
scipy
scikit-learn
matplotlib